refactor: deduplicate escape_atom_chars helper (BT-2113)
Move `escape_atom_chars` from `selector_mangler.rs` to the shared
`util.rs` and re-export it from `core_erlang`. Delete the identical
`escape_erlang_atom` copy in `beamtalk-cli/repl_startup.rs`; update
the one call site to use the canonical version.
