+++ crates/beamtalk-core/src/codegen/core_erlang/mod.rs
@@ -107,6 +107,7 @@ mod value_type_codegen;
 mod variable_context;
 
 // Re-export utility functions for IDE queries
+pub use util::escape_atom_chars;
 pub use util::to_module_name;
 
 use crate::ast::{Block, Expression, MessageSelector, Module, WellKnownSelector};
